pyproject.toml
[project]
name = "sentinel-sat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datetime>=6.0",
    "geopandas>=1.1.3",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "plotly>=6.7.0",
    "pystac-client>=0.9.0",
    "python-dotenv>=1.2.2",
    "rasterio>=1.5.0",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "sentinelhub>=3.11.5",
    "shapely>=2.1.2",
    "pytest>=8.4.2",
]

[project.scripts]
sentinel-demo = "scripts.run_sentinel_demo:main"
